loading of inbox changed
