+++ K6/browser/arbeidsflate/open-alle-virksomheter.js
@@ -8,7 +8,7 @@
 import { browser } from "k6/browser";
 import { check } from "k6";
 import { Trend } from "k6/metrics";
-import { getCookie, afUrl, environment } from "./arbeidsflate-utils.js";
+import { getCookie, afUrl, environment, waitForPageLoaded } from "./arbeidsflate-utils.js";
 
 const pageLoadingTime = new Trend("page_loading_time", true);
 const allOrganizationsTime = new Trend("all_organizations_time", true);
@@ -163,22 +163,4 @@ export async function selectAllEnterprises(page, trend, labels) {
     await waitForPageLoaded(page, 2);
     const endTime = new Date();
     trend.add(endTime - startTime, labels);
-}
-
-/**
- * Async function to wait for the page to load.
- * @param {object} page - The page object to interact with.
- * @param {number} empties - Number of empty checks to perform (default is 1).
- * @return {Promise<void>} - A promise that resolves when the page is loaded.
- */
-export async function waitForPageLoaded(page, empties = 1) {
-    let busyItems = await page.$$("li [aria-busy=\"true\"]");
-    let noEmptys = 0;
-    while (busyItems.length > 0 || noEmptys < empties) {
-        await page.waitForTimeout(10); // Wait for 10 ms before checking again
-        busyItems = await page.$$("li [aria-busy=\"true\"]");
-        if (busyItems.length === 0) {
-            noEmptys++;
-        }
-    }
 }
